+++ src/ai/backend/logging/handler/logstash.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import json
 import logging
 import socket
 import ssl
